opprydding. Fjern ubrukt kode, unødvendige rader. Omstrukturer
