+++ pom.xml
@@ -23,7 +23,7 @@
         <mockk.version>1.9.3</mockk.version>
         <confluent.version>5.3.1</confluent.version>
         <felles.version>1.20200624154125_0d8ff39</felles.version>
-        <kontrakter.version>2.0_20200625154602_9f6d526</kontrakter.version>
+        <kontrakter.version>2.0_20200716100404_0e17d4f</kontrakter.version>
         <main-class>no.nav.familie.ba.mottak.LauncherKt</main-class>
         <spring.vault.version>2.2.2.RELEASE</spring.vault.version>
         <spring.cloud.version>2.2.3.RELEASE</spring.cloud.version>
@@ -119,7 +119,7 @@
         <dependency>
             <groupId>no.nav.familie.kontrakter</groupId>
             <artifactId>barnetrygd</artifactId>
-            <version>2.0_20200710103407_aee01b8</version>
+            <version>${kontrakter.version}</version>
         </dependency>
 
         <!-- Kafka -->
