--- src/main/kotlin/no/nav/familie/ba/mottak/repository/SoknadRepository.kt
@@ -1,8 +1,0 @@
-package no.nav.familie.ba.mottak.repository
-
-import no.nav.familie.ba.mottak.repository.domain.Soknad
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface SoknadRepository : JpaRepository<Soknad, String>
