+++ src/main/resources/db/migration/V7__soknad.sql
@@ -1,7 +1,8 @@
 CREATE TABLE IF NOT EXISTS soknad(
-  id             UUID NOT NULL PRIMARY KEY, --UUID eller VARCHAR? Må UUID settes manuelt?
-  soknad_json    BYTEA NOT NULL, -- BYTEA og ikke json(b)
-  fnr            VARCHAR(50), -- Hvorfor 50?
-  task_opprettet BOOLEAN DEFAULT FALSE NOT NULL, --trenger vi den?
+  id             bigint NOT NULL PRIMARY KEY,
+  soknad_json    TEXT NOT NULL,
+  fnr            VARCHAR,
   opprettet_tid  TIMESTAMP(3) DEFAULT LOCALTIMESTAMP NOT NULL
 )
+
+CREATE SEQUENCE soknad_seq INCREMENT BY 50 START WITH 1000000 NO CYCLE;
